--- app/src/main/java/ru/practicum/android/diploma/data/sharedprefs/SharedPrefsStorageFilters.kt
@@ -1,11 +1,0 @@
-package ru.practicum.android.diploma.data.sharedprefs
-
-import ru.practicum.android.diploma.data.dto.SaveFiltersSharedPrefsDto
-
-interface SharedPrefsStorageFilters {
-    suspend fun read(): SaveFiltersSharedPrefsDto?
-
-    suspend fun write(filters: SaveFiltersSharedPrefsDto)
-
-    suspend fun clear()
-}
