+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -1,13 +1,12 @@
 package ru.practicum.android.diploma.di
 
+import com.google.gson.Gson
 import org.koin.dsl.module
 import ru.practicum.android.diploma.data.impl.FavouriteVacanciesRepositoryImpl
 import ru.practicum.android.diploma.data.impl.FilterRepositoryImpl
 import ru.practicum.android.diploma.data.impl.SearchRepositoryImpl
-import ru.practicum.android.diploma.data.impl.SharedPrefsRepositoryImpl
 import ru.practicum.android.diploma.data.impl.VacancyRepositoryImpl
 import ru.practicum.android.diploma.domain.FavouriteVacanciesRepository
-import ru.practicum.android.diploma.domain.SharedPrefsRepository
 import ru.practicum.android.diploma.domain.api.FilterRepository
 import ru.practicum.android.diploma.domain.api.SearchRepository
 import ru.practicum.android.diploma.domain.api.VacancyRepository
@@ -28,9 +27,12 @@ val repositoryModule = module {
         VacancyRepositoryImpl(get(), get())
     }
     single<FilterRepository> {
-        FilterRepositoryImpl(networkClient = get())
-    }
-    single<SharedPrefsRepository> {
-        SharedPrefsRepositoryImpl(get())
+        FilterRepositoryImpl(
+            networkClient = get(),
+            gson = provideGson(),
+            sharedPreferences = get()
+        )
     }
 }
+
+private fun provideGson() = Gson()
