+++ app/src/main/java/ru/practicum/android/diploma/domain/FilterInteractor.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.flow.Flow
 import ru.practicum.android.diploma.domain.models.Country
 import ru.practicum.android.diploma.domain.models.Industries
 import ru.practicum.android.diploma.domain.models.Region
+import ru.practicum.android.diploma.domain.models.SaveFiltersSharedPrefs
 import ru.practicum.android.diploma.util.ResponseData
 
 interface FilterInteractor {
@@ -11,4 +12,7 @@ interface FilterInteractor {
     suspend fun getRegions(id: String): Flow<ResponseData<List<Region>>>
     suspend fun getAllRegions(): Flow<ResponseData<List<Region>>>
     suspend fun getIndustries(): Flow<ResponseData<List<Industries>>>
+    suspend fun readSharedPrefs(): SaveFiltersSharedPrefs?
+    suspend fun writeSharedPrefs(filters: SaveFiltersSharedPrefs)
+    suspend fun clearSharedPrefs()
 }
