--- app/src/main/java/ru/practicum/android/diploma/domain/SharedPrefsInteractor.kt
@@ -1,11 +1,0 @@
-package ru.practicum.android.diploma.domain
-
-import ru.practicum.android.diploma.domain.models.SaveFiltersSharedPrefs
-
-interface SharedPrefsInteractor {
-    suspend fun read(): SaveFiltersSharedPrefs?
-
-    suspend fun write(filters: SaveFiltersSharedPrefs)
-
-    suspend fun clear()
-}
