--- app/src/main/java/ru/practicum/android/diploma/domain/SharedPrefsRepository.kt
@@ -1,11 +1,0 @@
-package ru.practicum.android.diploma.domain
-
-import ru.practicum.android.diploma.domain.models.SaveFiltersSharedPrefs
-
-interface SharedPrefsRepository {
-    suspend fun read(): SaveFiltersSharedPrefs?
-
-    suspend fun write(filters: SaveFiltersSharedPrefs)
-
-    suspend fun clear()
-}
