+++ app/src/main/java/ru/practicum/android/diploma/domain/api/FilterRepository.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.flow.Flow
 import ru.practicum.android.diploma.domain.models.Country
 import ru.practicum.android.diploma.domain.models.Industries
 import ru.practicum.android.diploma.domain.models.Region
+import ru.practicum.android.diploma.domain.models.SaveFiltersSharedPrefs
 import ru.practicum.android.diploma.util.ResponseData
 
 interface FilterRepository {
@@ -11,4 +12,7 @@ interface FilterRepository {
     fun getRegions(id: String): Flow<ResponseData<List<Region>>>
     fun getAllRegions(): Flow<ResponseData<List<Region>>>
     fun getIndustries(): Flow<ResponseData<List<Industries>>>
+    suspend fun readSharedPrefs(): SaveFiltersSharedPrefs?
+    suspend fun writeSharedPrefs(filters: SaveFiltersSharedPrefs)
+    suspend fun clearSharedPrefs()
 }
